refactor: improve mock data generation and remove unnecessary transaction
This commit refactors the mock data generation process and removes an unnecessary database transaction during port scan data parsing. It also includes minor updates to logging and JSON formatting. The mock data generation is now more context-aware and handles different task types more effectively. Removing the transaction improves the performance of port data processing. These changes enhance the efficiency and maintainability of the codebase.
